refactor(frontend): MkHorizontalSwipe -> MkSwiper
